Reorganize Home route
