+++ src/components/Editor/ReactFlow/FlowSidebar/components/UserComponentItem.tsx
@@ -2,7 +2,7 @@ import { useQueryClient } from "@tanstack/react-query";
 import { File, Trash2 } from "lucide-react";
 import { type DragEvent, useCallback, useMemo } from "react";
 
-import ConfirmationDialog from "@/components/ConfirmationDialog";
+import ConfirmationDialog from "@/components/shared/ConfirmationDialog";
 import { SidebarMenuItem } from "@/components/ui/sidebar";
 import {
   Tooltip,
