+++ src/components/Home/PipelineSection/PipelineRow/index.ts
