+++ src/components/Home/PipelineSection/index.ts
