+++ src/components/shared/Status/types.ts
@@ -1,0 +1,8 @@
+export interface TaskStatusCounts {
+  total: number;
+  succeeded: number;
+  failed: number;
+  running: number;
+  waiting: number;
+  skipped: number;
+}
